Remove situation in the back
