+++ backend/models/answer.js
@@ -1,4 +1,11 @@
 var mongoose = require("mongoose")
+var utils = require("../lib/utils")
+var openfisca = require("../lib/openfisca")
+var mesAides = require("../lib/mes-aides")
+var benefits = require("../../app/js/constants/benefits/back")
+const { generateSituation } = require("../../lib/situations")
+
+var computeAides = mesAides.computeAides.bind(benefits)
 
 const answer = {
   entityName: String,
@@ -16,8 +23,54 @@ var AnswerSchema = new mongoose.Schema(
     patrimoine: Object,
     dateDeValeur: Date,
     version: Number,
+    abtesting: Object,
+    createdAt: { type: Date, default: Date.now },
+    modifiedFrom: String,
+    status: {
+      type: String,
+      default: "new",
+      enum: ["new", "test", "investigation"],
+    },
+    token: String,
   },
   { minimize: false }
 )
+
+AnswerSchema.statics.cookiePrefix = "situation_"
+AnswerSchema.virtual("cookieName").get(function () {
+  return `${AnswerSchema.statics.cookiePrefix}${this._id}`
+})
+AnswerSchema.methods.isAccessible = function (keychain) {
+  return (
+    ["demo", "investigation", "test"].includes(this.status) ||
+    (keychain && keychain[this.cookieName] === this.token)
+  )
+}
+AnswerSchema.pre("save", function (next) {
+  if (!this.isNew) {
+    return next()
+  }
+  var answers = this
+  utils
+    .generateToken()
+    .then(function (token) {
+      answers.token = token
+    })
+    .then(next)
+    .catch(next)
+})
+AnswerSchema.methods.compute = function () {
+  var situation = generateSituation(this)
+  return new Promise(function (resolve, reject) {
+    openfisca.calculate(situation, function (err, openfiscaResponse) {
+      if (err) {
+        return reject(err)
+      }
+
+      var aides = computeAides(situation, this._id, openfiscaResponse, false)
+      resolve(aides)
+    })
+  })
+}
 
 mongoose.model("Answer", AnswerSchema)
