+++ lib/constants/resources.js
@@ -1,5 +1,5 @@
 "use strict"
-const Individu = require("../lib/Individu")
+const Individu = require("../../src/lib/Individu")
 const datesGenerator = require("../../backend/lib/mes-aides/dates").generator
 
 let ressourceCategories = [
