+++ lib/ressource.js
@@ -1,8 +1,9 @@
-import { categoriesRnc, ressourceTypes } from "@/constants/resources"
-import filter from "lodash/filter"
-import keys from "lodash/keys"
-import keyBy from "lodash/keyBy"
-import uniq from "lodash/uniq"
+var axios = require("axios")
+var { categoriesRnc, ressourceTypes } = require("./constants/resources")
+var filter = require("lodash/filter")
+var keys = require("lodash/keys")
+var keyBy = require("lodash/keyBy")
+var uniq = require("lodash/uniq")
 
 function getPeriodsForCurrentYear(dates, ressourceType) {
   let periodKeys = []
@@ -171,6 +172,17 @@ function isRessourceOnMainScreen(ressourceOrType) {
   return type != "pensions_alimentaires_versees_individu"
 }
 
+function getParameterFromOpenfisca(parameterId) {
+  return axios
+    .get("/api/openfisca/parameters/" + parameterId)
+    .then(function (resp) {
+      let values = resp.data.values
+      let sortedByDates = Object.keys(values).sort()
+      let latestValue = values[sortedByDates.pop()]
+      return latestValue
+    })
+}
+
 const Ressource = {
   getPeriodsForCurrentYear,
   // Ne semble pas être utilisée
@@ -184,6 +196,7 @@ const Ressource = {
   getIndividuRessourceTypesByCategory,
   setIndividuRessourceTypes,
   unsetForCurrentYear,
+  getParameterFromOpenfisca,
 }
 
-export default Ressource
+module.exports = Ressource
