+++ src/views/Foyer/Ressources/Fiscales.vue
@@ -71,7 +71,7 @@ import sum from "lodash/sum"
 import some from "lodash/some"
 import isNaN from "lodash/isNaN"
 import Individu from "@/lib/Individu"
-import { categoriesRnc } from "@/constants/resources"
+import { categoriesRnc } from "../../../../lib/constants/resources"
 import Actions from "@/components/Actions"
 
 function getDefaultValue(months, individu, rnc) {
