+++ tests/unit/openfisca/test.spec.js
@@ -3,7 +3,7 @@ var { values } = require("lodash")
 var Promise = require("bluebird")
 var fs = Promise.promisifyAll(require("fs"))
 var subject = require("../../../backend/lib/openfisca/test")
-var resources = require("../../../src/constants/resources")
+var resources = require("../../../lib/constants/resources")
 var tmp = require("tmp")
 
 var details = {
